Implement bulk of Interpreter (UNTESTED) (what could go wrong)

+++ befunge/src/main/kotlin/befide/befunge/state/Value.kt
@@ -1,7 +1,7 @@
 package befide.befunge.state
 
-data class Value(val value: Int) {
-    constructor(value: Char) : this(value.toInt())
+data class Value(val value: Long) {
+    constructor(value: Char) : this(value.toLong())
 
     val asChar: Char? = if (value in (32..126)) value.toChar() else null
 }
